pyproject.toml
[tool.poetry]
name = "trucking-data"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "trucking_data"}]

[tool.poetry.dependencies]
python = "^3.11"
jupyterlab = "^4.4.7"
polars = "^1.33.1"
requests = "^2.32.5"
tqdm = "^4.67.1"
rapidfuzz = "^3.14.1"
numpy = "^2.3.3"
geopandas = "^1.1.1"
cartopy = "^0.25.0"
matplotlib = "^3.10.6"
pyarrow = "^21.0.0"
polars-strsim = "^0.2.5"
altair = "^5.5.0"
vegafusion = {version = ">=1.5.0", extras = ["embed"]}
vl-convert-python = ">=1.6.0"
dnspython = "^2.8.0"
plotly = "^6.3.1"
seaborn = "^0.13.2"
dotenv = "^0.9.9"
openai = "^2.8.1"
sodapy = "^2.2.0"
google-cloud-aiplatform = "^1.127.0"
ipykernel = "^7.1.0"
scikit-learn = "^1.7.2"
streamlit = "^1.51.0"
xgboost = "^3.1.1"
joypy = "^0.2.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.main]
init-hook = "import sys; sys.path.insert(0, '.')"
